Controller 분리 및 구독-발행 패턴 적용
